enable support for userland promises
