+++ src/utils.js
@@ -1,4 +1,11 @@
-import { extractUrlAndGlobal, customPromise as customPromiseTemplate } from './internal';
+// split the @ syntax into url and global
+export const extractUrlAndGlobal = (urlAndGlobal) => {
+  const index = urlAndGlobal.indexOf('@');
+  if (index <= 0 || index === urlAndGlobal.length - 1) {
+    throw new Error(`Invalid request "${urlAndGlobal}"`);
+  }
+  return [urlAndGlobal.substring(index + 1), urlAndGlobal.substring(0, index)];
+};
 
 const remoteVars = process.env.REMOTES || {};
 
@@ -108,5 +115,3 @@ export const injectScript = (keyOrRuntimeRemoteItem) => {
       return container;
     });
 };
-
-export const customPromise = customPromiseTemplate
